+++ bot/application/student/exception/invalid_discord_id_format_exception.py
@@ -1,4 +1,4 @@
-from bot.application.student.exceptions.invalid_format_exception import (
+from bot.application.student.exception.invalid_format_exception import (
     InvalidFormatException,
 )
