+++ bot/application/student/exception/student_already_exist.py
