+++ bot/application/student/exception/student_already_registered_exception.py
