+++ bot/resource/cog/validator/name_validator.py
