+++ bot/resource/cog/validator/ni_validator.py
